test/requirements-test.txt
pytest==7.4.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0